Add files via upload
Adding the first to sage scripts that assisted me with tutorial 3 to verify my answers, I just simply use "calc_ff_elem" to make sure my addition and multiplication operations are correct in the finite field "ff". "matrix_echlon_ff" is a bit more involved; this shows the finite field elements,  shows their addition and multiplication tables, and takes a matrix of elements from the finite field and reduces it.